+++ packages/backend/src/mobileBackendManager.ts
@@ -1,6 +1,8 @@
-import { DataServer } from './socket/DataServer'
-import { ConnectionsManager } from './libp2p/connectionsManager'
 import { Command } from 'commander'
+import { ConnectionsManager } from './libp2p/connectionsManager'
+
+import logger from './logger'
+const log = logger('conn')
 
 export const runBackend = async (): Promise<any> => {
   // Enable triggering push notifications
@@ -8,17 +10,16 @@ export const runBackend = async (): Promise<any> => {
   process.env['CONNECTION_TIME'] = (new Date().getTime() / 1000).toString() // Get time in seconds
 
   const program = new Command()
-
+  
   program
-    .requiredOption('-d, --appDataPath <appDataPath>', 'app data path')
-    .requiredOption('-p, --dataPort <dataPort>', 'data port')
-    .requiredOption('-t, --httpTunnelPort <httpTunnelPort>', 'httpTunnelPort')
-    .requiredOption('-s, --socksPort <socksPort>', 'socks port')
-    .requiredOption('-c, --controlPort <controlPort>', 'control port')
-    .requiredOption('-a, --authCookie <authCookie>', 'control port authentication cookie')
-
+  .requiredOption('-d, --appDataPath <appDataPath>', 'app data path')
+  .requiredOption('-p, --dataPort <dataPort>', 'data port')
+  .requiredOption('-t, --httpTunnelPort <httpTunnelPort>', 'httpTunnelPort')
+  .requiredOption('-s, --socksPort <socksPort>', 'socks port')
+  .requiredOption('-c, --controlPort <controlPort>', 'control port')
+  .requiredOption('-a, --torPath <torPath>', 'tor binary path')
+  
   program.parse(process.argv)
-
   const options = program.opts()
 
   const connectionsManager: ConnectionsManager = new ConnectionsManager({
@@ -27,7 +28,8 @@ export const runBackend = async (): Promise<any> => {
     socketIOPort: options.dataPort,
     options: {
       env: {
-        appDataPath: options.appDataPath
+        appDataPath: options.appDataPath,
+        resourcesPath: options.torPath
       },
       createPaths: false,
       torControlPort: options.controlPort,