+++ packages/backend/src/storage/storage.ts
@@ -166,7 +166,7 @@ export class Storage {
 
   protected async initIPFS(libp2p: Libp2p, peerID: PeerId): Promise<IPFS.IPFS> {
     log('Initializing IPFS')
-    // @ts-expect-error
+    // @ts-ignore
     return await IPFS.create({
       // error here 'permission denied 0.0.0.0:443'
       libp2p: async () => libp2p,