+++ packages/mobile/android/app/src/main/java/com/zbaymobile/MainActivity.java
@@ -1,7 +1,9 @@
 package com.zbaymobile;
 
+import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.modules.core.DeviceEventManagerModule;
@@ -18,16 +20,18 @@ public class MainActivity extends ReactActivity {
         return "ZbayMobile";
     }
 
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(null);
-    }
-
     private void sendNotificationInfo (ReactApplicationContext reactContext, Intent intent) {
         String channelAddress = intent.getStringExtra("channelAddress");
         if (channelAddress != null) {
             reactContext.getJSModule(DeviceEventManagerModule.RCTDeviceEventEmitter.class).emit("notification", channelAddress);
         }
+    }
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(null);
+        Context context = getApplicationContext();
+        new BackendWorkManager(context).enqueueRequests();
     }
 
     @Override
@@ -41,6 +45,12 @@ public class MainActivity extends ReactActivity {
         if (tag.equals("notification")) {
             ReactApplicationContext reactContext = (ReactApplicationContext) getReactNativeHost().getReactInstanceManager().getCurrentReactContext();
             this.sendNotificationInfo(reactContext, intent);
-        };
+        }
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        Log.d("QUIET", "Application destroyed.");
     }
 }