+++ packages/mobile/android/app/src/main/java/com/zbaymobile/Utils/Const.kt
@@ -1,7 +1,5 @@
 package com.zbaymobile.Utils
 
-import com.zbaymobile.MainApplication.PACKAGE_NAME
-
 object Const {
     const val TAG_NOTICE="NOTICE"
 
@@ -9,8 +7,6 @@ object Const {
     const val TAG_TOR_ERR = "TOR_ERR"
 
     const val INCOMING_MESSAGES_CHANNEL_ID = "incoming.messages.channel"
-
-    val SERVICE_ACTION_EXECUTE = "$PACKAGE_NAME.service.execute"
 
     const val TOR_LOGS_FILENAME = "tor.log"
 }