+++ packages/mobile/src/store/init/init.master.saga.ts
@@ -1,11 +1,11 @@
 import { all, takeEvery } from 'typed-redux-saga'
-import { doOnRestoreSaga } from './doOnRestore/doOnRestore.saga'
 import { initActions } from './init.slice'
+import { onRestoreSaga } from './onRestore/onRestore.saga'
 import { startConnectionSaga } from './startConnection/startConnection.saga'
 
 export function* initMasterSaga(): Generator {
   yield all([
-    takeEvery(initActions.doOnRestore.type, doOnRestoreSaga),
-    takeEvery(initActions.onBackendStarted.type, startConnectionSaga)
+    takeEvery(initActions.onRestore.type, onRestoreSaga),
+    takeEvery(initActions.startWebsocketConnection.type, startConnectionSaga)
   ])
 }