+++ packages/mobile/src/store/nativeServices/events/nativeEvent.keys.ts
@@ -1,0 +1,4 @@
+export enum NativeEventKeys {
+  Backend = 'backend',
+  Notification = 'notification'
+}