+++ packages/identity/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.3-alpha.16](https://github.com/TryQuiet/identity/compare/@quiet/identity@0.10.3-alpha.13...@quiet/identity@0.10.3-alpha.16) (2022-10-19)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+## [0.10.3-alpha.15](https://github.com/TryQuiet/identity/compare/@quiet/identity@0.10.3-alpha.13...@quiet/identity@0.10.3-alpha.15) (2022-10-18)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+## [0.10.3-alpha.14](https://github.com/TryQuiet/identity/compare/@quiet/identity@0.10.3-alpha.13...@quiet/identity@0.10.3-alpha.14) (2022-10-18)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
 ## [0.10.3-alpha.13](https://github.com/TryQuiet/identity/compare/@quiet/identity@0.10.3-alpha.7...@quiet/identity@0.10.3-alpha.13) (2022-09-22)
 
 **Note:** Version bump only for package @quiet/identity